allow using other storage mechanisms for attachments than gridstore
