+++ api.js
@@ -104,8 +104,9 @@ module.exports = done => {
     messageHandler = new MessageHandler({
         database: db.database,
         users: db.users,
+        redis: db.redis,
         gridfs: db.gridfs,
-        redis: db.redis
+        attachments: config.attachments
     });
 
     userHandler = new UserHandler({
