+++ lib/attachment-storage.js
@@ -1,8 +1,7 @@
 'use strict';
 
-const ObjectID = require('mongodb').ObjectID;
+const GridstoreStorage = require('./attachments/gridstore-storage.js');
 const crypto = require('crypto');
-const GridFSBucket = require('mongodb').GridFSBucket;
 let cryptoAsync;
 try {
     cryptoAsync = require('@ronomon/crypto-async'); // eslint-disable-line global-require
@@ -12,30 +11,20 @@ try {
 
 class AttachmentStorage {
     constructor(options) {
-        this.bucketName = options.bucket || 'attachments';
-        this.gridfs = options.gridfs;
-        this.gridstore = new GridFSBucket(this.gridfs, {
-            bucketName: this.bucketName
-        });
+        this.options = options || {};
+
+        let type = (options.options && options.options.type) || 'gridstore';
+
+        switch (type) {
+            case 'gridstore':
+            default:
+                this.storage = new GridstoreStorage(this.options);
+                break;
+        }
     }
 
     get(attachmentId, callback) {
-        this.gridfs.collection('attachments.files').findOne({
-            _id: attachmentId
-        }, (err, attachmentData) => {
-            if (err) {
-                return callback(err);
-            }
-            if (!attachmentData) {
-                return callback(new Error('This attachment does not exist'));
-            }
-
-            return callback(null, {
-                contentType: attachmentData.contentType,
-                transferEncoding: attachmentData.metadata.transferEncoding,
-                metadata: attachmentData.metadata
-            });
-        });
+        return this.storage.get(attachmentId, callback);
     }
 
     create(attachment, callback) {
@@ -43,68 +32,12 @@ class AttachmentStorage {
             if (err) {
                 return callback(err);
             }
-
-            this.gridfs.collection(this.bucketName + '.files').findOneAndUpdate({
-                'metadata.h': hash
-            }, {
-                $inc: {
-                    'metadata.c': 1,
-                    'metadata.m': attachment.magic
-                }
-            }, {
-                returnOriginal: false
-            }, (err, result) => {
-                if (err) {
-                    return callback(err);
-                }
-
-                if (result && result.value) {
-                    return callback(null, result.value._id);
-                }
-
-                let returned = false;
-
-                let id = new ObjectID();
-                let metadata = {
-                    h: hash,
-                    m: attachment.magic,
-                    c: 1,
-                    transferEncoding: attachment.transferEncoding
-                };
-                Object.keys(attachment.metadata || {}).forEach(key => {
-                    if (!(key in attachment.metadata)) {
-                        metadata[key] = attachment.metadata[key];
-                    }
-                });
-
-                let store = this.gridstore.openUploadStreamWithId(id, null, {
-                    contentType: attachment.contentType,
-                    metadata
-                });
-
-                store.once('error', err => {
-                    if (returned) {
-                        return;
-                    }
-                    returned = true;
-                    callback(err);
-                });
-
-                store.once('finish', () => {
-                    if (returned) {
-                        return;
-                    }
-                    returned = true;
-                    return callback(null, id);
-                });
-
-                store.end(attachment.body);
-            });
+            return this.storage.create(attachment, hash, callback);
         });
     }
 
     createReadStream(id) {
-        return this.gridstore.openDownloadStream(id);
+        return this.storage.createReadStream(id);
     }
 
     deleteMany(ids, magic, callback) {
@@ -120,57 +53,15 @@ class AttachmentStorage {
     }
 
     updateMany(ids, count, magic, callback) {
-        // update attachments
-        this.gridfs.collection(this.bucketName + '.files').updateMany(
-            {
-                _id: {
-                    $in: ids
-                }
-            },
-            {
-                $inc: {
-                    'metadata.c': count,
-                    'metadata.m': magic
-                }
-            },
-            {
-                multi: true,
-                w: 1
-            },
-            callback
-        );
+        this.storage.update(ids, count, magic, callback);
     }
 
     delete(id, magic, callback) {
-        this.gridfs.collection(this.bucketName + '.files').findOneAndUpdate({
-            _id: id
-        }, {
-            $inc: {
-                'metadata.c': -1,
-                'metadata.m': -magic
-            }
-        }, {
-            returnOriginal: false
-        }, (err, result) => {
-            if (err) {
-                return callback(err);
-            }
+        this.storage.delete(id, magic, callback);
+    }
 
-            if (!result || !result.value) {
-                return callback(null, false);
-            }
-
-            if (result.value.metadata.c === 0 && result.value.metadata.m === 0) {
-                return this.gridstore.delete(id, err => {
-                    if (err) {
-                        return callback(err);
-                    }
-                    callback(null, 1);
-                });
-            }
-
-            return callback(null, 0);
-        });
+    deleteOrphaned(callback) {
+        this.storage.deleteOrphaned(callback);
     }
 
     calculateHash(input, callback) {
