+++ lib/message-handler.js
@@ -26,7 +26,8 @@ class MessageHandler {
         this.attachmentStorage =
             options.attachmentStorage ||
             new AttachmentStorage({
-                gridfs: options.gridfs || options.database
+                gridfs: options.gridfs || options.database,
+                options: options.attachments
             });
 
         this.indexer = new Indexer({
